LinkedIn Auto Endorse DEV: issue #25
This commit provides a way to face with the problem descripted in the issue.

In some profiles, it looks like that the skills section wasn't properly loaded. LinkedIn doesn't entirely load all sections content when opening the profile.
If there are some loading spinners in the page, the script will jump to all spinners.