LinkedIn Profile Scrapper: issue #35
The TypeError randomly comes when a section in the page isn't fully loaded.
By slowing down the function the error doesn't seem to appear.

When debugging the script some profiles seems to not be loaded on getProfileInfos, now adding for the waitUntilVisible 7500ms for timeout.

Additionnaly new skills selectors are added.
If the skills selector doesn't return skills items, the script will try to use alternative selectors to scrape skills.